### 2022-06-22 #4499 @Kevin-Mattheus-Moerman
@btmonier thanks for this JOSS submission. I am the AEiC on call this week and here to help with the initial steps. I have started searching for a handling editor and invited @fboehm as suggested.
Note that our system flagged some potentially missing DOIs :point_up: , can you check these? You can call `@editorialbot generate pdf` to update the paper, and you can call `@editorialbot check references` to re-check the DOIs too. Thanks.
