### 2022-06-22 #4499 @fboehm
@Tgerke, @jhollist - would you be willing to review this submission for JOSS?
