### 2022-06-23 e8c8f3f
Add missing DOI; remove Zotero file locations
