+++ paper/paper.bib
@@ -15,7 +15,6 @@
 	year = {2007},
 	note = {Publisher: Oxford Academic},
 	pages = {2633--2635},
-	file = {Full Text PDF:C\:\\Users\\brand\\Zotero\\storage\\NL2YLJ98\\Bradbury et al. - 2007 - TASSEL software for association mapping of comple.pdf:application/pdf;Snapshot:C\:\\Users\\brand\\Zotero\\storage\\DA3UN2DZ\\185151.html:text/html},
 }
 
 @article{zhang_software_2009,
@@ -34,7 +33,6 @@
 	year = {2009},
 	note = {Publisher: Oxford Academic},
 	pages = {664--675},
-	file = {Full Text PDF:C\:\\Users\\brand\\Zotero\\storage\\4UJ5NES3\\Zhang et al. - 2009 - Software engineering the mixed model for genome-wi.pdf:application/pdf;Snapshot:C\:\\Users\\brand\\Zotero\\storage\\6LNS3A78\\260106.html:text/html},
 }
 
 
@@ -54,7 +52,6 @@
 	month = sep,
 	year = {2004},
 	pages = {R80},
-	file = {Springer Full Text PDF:C\:\\Users\\brand\\Zotero\\storage\\F5NXDS2C\\Gentleman et al. - 2004 - Bioconductor open software development for comput.pdf:application/pdf},
 }
 
 @article{lawrence_software_2013,
@@ -74,7 +71,6 @@
 	note = {Publisher: Public Library of Science},
 	keywords = {Sequence alignment, Genome analysis, Genome annotation, Genome complexity, Genomic databases, Mammalian genomics, Software tools, Structural genomics},
 	pages = {e1003118},
-	file = {Full Text PDF:C\:\\Users\\brand\\Zotero\\storage\\C5Z4I2L7\\Lawrence et al. - 2013 - Software for Computing and Annotating Genomic Rang.pdf:application/pdf;Snapshot:C\:\\Users\\brand\\Zotero\\storage\\MTWC6HAL\\article.html:text/html},
 }
 
 @software{morgan_summarizedexperiment_2020,
@@ -111,7 +107,6 @@
 	note = {Number: 2
 Publisher: Nature Publishing Group},
 	pages = {203--208},
-	file = {Full Text PDF:C\:\\Users\\brand\\Zotero\\storage\\3SMVS52Q\\Yu et al. - 2006 - A unified mixed-model method for association mappi.pdf:application/pdf;Snapshot:C\:\\Users\\brand\\Zotero\\storage\\ZEVDIQDZ\\ng1702.html:text/html},
 }
 
 @article{shabalin_matrix_2012,
@@ -131,7 +126,6 @@ Publisher: Nature Publishing Group},
 	year = {2012},
 	note = {Publisher: Oxford Academic},
 	pages = {1353--1358},
-	file = {Full Text PDF:C\:\\Users\\brand\\Zotero\\storage\\G94JK79G\\Shabalin - 2012 - Matrix eQTL ultra fast eQTL analysis via large ma.pdf:application/pdf;Snapshot:C\:\\Users\\brand\\Zotero\\storage\\FXKBEH3G\\213326.html:text/html},
 }
 
 @book{wickham_ggplot2_2016,
@@ -163,7 +157,6 @@ Publisher: Nature Publishing Group},
 Section: Genomic Selection},
 	keywords = {genomic selection, GenPred, breeding value prediction, realized relationship matrix, Shared Data Resources, shrinkage estimation},
 	pages = {1405--1413},
[…]
