### 2022-06-26 #4499 @stevokeyo
Hello,

Unfortunately I will nor be in a position to take-up this review seeing its scope is out of my area of specialisation.

Stephen.

> On 24 Jun 2022, at 21:11, Frederick Boehm ***@***.***> wrote:
>
>
> @simonpcouch <https://github.com/simonpcouch> @flor14 <https://github.com/flor14> @bioedd <https://github.com/bioedd> @stevokeyo <https://github.com/stevokeyo> @brunomontezano <https://github.com/brunomontezano> @stmcg <https://github.com/stmcg> - Would you be willing to review this submission for JOSS?
>
