### 2022-06-27 #4499 @fboehm
@S-canchi @petermacp @stensy @darokun @monicagerber @darrendahly - would you be willing to review this submission for JOSS? Given your expertise in R, I think that you would make valuable contributions to the review. Here is the link to the package: https://github.com/maize-genetics/rTASSEL. At JOSS, our reviews are driven by checklists.
