### 2022-06-27 #4499 @bioedd
Hello. Thanks for the invite! Unfortunately, I won't be able to review as
this submission is outside of my area of expertise.

Eduardo Álvarez
Editor Asociado - Economía Creativa

Av. Constituyentes 455, Col. América
[55] 2789 9000 | centro.edu.mx <http://www3.centro.edu.mx/> | @centro_u
<http://www.instagram.com/centro_u>

El vie, 24 jun 2022 a las 13:11, Frederick Boehm ***@***.***>)
escribió:

> @simonpcouch <https://github.com/simonpcouch> @flor14
> <https://github.com/flor14> @bioedd <https://github.com/bioedd> @stevokeyo
> <https://github.com/stevokeyo> @brunomontezano
> <https://github.com/brunomontezano> @stmcg <https://github.com/stmcg> -
> Would you be willing to review this submission for JOSS?
>
