### 2022-06-28 #4499 @petermacp
Many thanks for asking me to review, but also unfortunately not in an area with which I have expertise.

All the best,

Peter

Dr Peter MacPherson MBChB MPH PhD
Wellcome Trust Fellow & Reader in Population Health
Liverpool School of Tropical Medicine &
Malawi-Liverpool-Wellcome Clinical Research Programme, Blantyre, Malawi

Honorary Professor in Population Health
London School of Hygiene and Tropical Medicine

Honorary Consultant in Communicable Disease Control
UK Health Security Agency

Mob: +265 99 717 6230
Email: ***@***.*** ***@***.***>
Website: Website: https://petermacpherson.netlify.app/ <https://petermacpherson.netlify.app/>

> On 28 Jun 2022, at 00:21, Frederick Boehm ***@***.***> wrote:
>
>
> Thanks for the reply, @bioedd <https://github.com/bioedd> . I'll keep you in mind for a future submission.
>
