### 2022-06-29 #4499 @flor14
Thanks for the invite. I am willing to review but this is outside of my
expertise.
Best,
Florencia

El vie, 24 jun 2022 a las 12:33, The Open Journals editorial robot (<
***@***.***>) escribió:

> 👉📄 Download article proof
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.04499/joss.04499/10.21105.joss.04499.pdf>
> 📄 View article proof on GitHub
> <https://github.com/openjournals/joss-papers/blob/joss.04499/joss.04499/10.21105.joss.04499.pdf>
> 📄 👈
>
