### 2022-06-29 #4530 @editorialbot
[REVIEW]: rTASSEL: An R interface to TASSEL for analyzing genomic diversity
**Submitting author:** @btmonier (<a href="http://orcid.org/0000-0001-6797-1221">Brandon Monier</a>)
**Repository:** https://github.com/maize-genetics/rTASSEL
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.9.27
