### 2022-07-03 maize-genetics/rTASSEL#3 @tomsing1
[JOSS review] Add `dependencies = TRUE` to build vignettes on a fresh system.
Congratulations on this awesome R package! As part of the JOSS review, I have installed the package within a Bioconductor docker image (R 4.2.0). I noticed that the

```r
devtools::install_bitbucket(
    repo = "bucklerlab/rtassel",
    ref = "master",
    build_vignettes = TRUE
)
```

command fails, because a fresh container doesn't have the `rmarkdown` package (listed under `Suggests` in rTASSEL's DESCRIPTION file) is not available.

To install packages listed under `Suggests`, users need to add `dependencies = TRUE` to the `devtools::install_bitbucket` command. Perhaps you want to this argument to make sure the `build_vignettes = TRUE` argument succeeds even on a new system?

https://github.com/maize-genetics/rTASSEL/blob/679c21d0ded09893ebabaa0b74e39ea61e27a873/vignettes/rtassel_walkthrough.Rmd#L122
