### 2022-07-03 maize-genetics/rTASSEL#3 @tomsing1
P.S.: You might also want to add this in the https://github.com/maize-genetics/rTASSEL#usage `Usage` section of the README.md file, because it explicitly mentions the `build_vignettes` parameter.
