### 2022-07-03 maize-genetics/rTASSEL#4 @tomsing1
[JOSS review]: Enriching pkgdown project site
Just a suggestion, but you might want to consider add the following to the (currently empty) `_pkgdown.yml` file:

```YAML
template:
  bootstrap: 5
url: https://maizee-genetics.github.io/rTASSEL/index.html
home:
  links:
  - text: Bitbucket
    href: https://bitbucket.org/bucklerlab/rtassel/src/master/
  - text: github
    href: https://github.com/maize-genetics/rTASSEL
articles:
- title: Articles
  navbar: ~
  contents:
  - rtassel_walkthrough
  - genotype_filtration
  - rtassel_benchmarks
```

to

1. Explicitly order the articles, e.g. putting the walkthrough first
2. Add links to the github and bitbucket repositories to the project site
3. Take advantage of bootstrap 5 (introduced with [pkgdown v. 2.0.0](https://www.tidyverse.org/blog/2021/12/pkgdown-2-0-0/))
