### 2022-07-03 maize-genetics/rTASSEL#5 @tomsing1
[JOSS review]: add contributing.md
Please add a `contributing.md` file to fulfill [JOSS' Community guidelines
](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines)

For inspiration, you might enjoy reading [mozilla's advice](https://mozillascience.github.io/working-open-workshop/contributing/). You might also consider running the `usethis::use_tidy_contributing()` and `usethis::use_tidy_coc()` functions from the [usethis R package](https://usethis.r-lib.org/) and modifying the auto-generated `.github/CONTRIBUTING.md`. and `.github/CODE_OF_CONDUCT.md` files.
