### 2022-07-03 maize-genetics/rTASSEL#6 @tomsing1
log4j:WARN Please initialize the log4j system properly.
It seems the log4j initialization might not be complete, as this code snipped from the `rtassel_walkthrough.Rmd` vignette gives a warning. (This is not relevant for the JOSS review, but I thought I'd point it out in case you want to check.)

```r
> library(rTASSEL)
Welcome to rTASSEL (version 0.9.26)
 • Consider starting a TASSEL log file (see ?startLogger())

> rTASSEL::startLogger(fullPath = NULL, fileName = NULL)
log4j:WARN No appenders could be found for logger (net.maizegenetics.util.LoggingUtils).
log4j:WARN Please initialize the log4j system properly.
TASSEL logging file created at: /home/rstudio/rTassel/vignettes/rTASSEL_log
```

https://github.com/maize-genetics/rTASSEL/blob/679c21d0ded09893ebabaa0b74e39ea61e27a873/R/LoggingSupport.R#L13
