### 2022-07-03 #4530 @tomsing1
@fboehm : The R package includes a vignette `rtassel_benchmarks.Rmd` that reports the results of an extensive benchmarking exercise of `rTASSEL'`s performance. I have neither the code, data nor the computational resources to reproduce these benchmarks.

Nevertheless, I think the value of this manuscript is clear, e.g. it is not (just) based on performance but by offering R users a robust way to interact with TASSEL. Hence I would like to tick the `Performance` box, taking the authors at their word. Any advice?
