### 2022-07-19 #4530 @fboehm
@btmonier - Thanks for your question. It's fine to add the changes to your master branch; you don't need to make a new branch. Also, adding the changes to the master branch will make it easier for the second reviewer to find them.
