### 2022-07-20 maize-genetics/rTASSEL#7 @tkchafin
[JOSS Review] Minor editorial comments
**Minor comments**

Nothing critical here, just a few suggestions -- consider giving the MS another run through to improve readability.

L12 — “TASSEL has limited means to use scripting languages” seems awkwardly worded. Maybe “has limited automation potential” or similar

L13 — Should there be a comma after “R package”?

L19 — The use of “evermore” seems awkward, consider replacing with a more commonly used phrasing

L46 — Remove comma after “function” and “readGenotypePhyenotype” or maybe revise L45-46 to “Once data is imported, the function readGenotypePhenotype is used to construct an S4 object, which is used for all downstream analyses”

L61-62 — Consider “Linear models can be specified following the format used by R’s lm function:”

L68 — Double citation

L88 — Might specify what sort of phylogenetic analysis this is (neighbor joining etc)
