### 2022-07-20 #4530 @tkchafin
@fboehm : I've completed my review. The submitted R package and accompanying manuscript are valuable contributions to the field, and adds considerable value (e.g., through integrated plotting and interaction with other R packages) beyond simply wrapping the TASSEL toolkit.

In addition to the issues pointed out by @tomsing1 (i.e., installation and contributing docs), I would suggest that the authors consider creating a containerized option, as I had to use some workarounds to get everything working on an M1 Mac. This would increase portability and save some headache for some users... But just a suggestion (I wouldn't consider this critical). I also suggested some very minor editorial changes, but again nothing critical.

Great job to the authors, and after fixing the minor issues which have been brought up in this review I would say that this has my full support for publication!
