### 2022-07-28 78683cd
Update vignette building for fresh systems
Issue #3
