### 2022-07-28 ca0b7d8
Enrich pkgdown project site
Issue #4
