### 2022-07-28 06572e8
Add contribution and code of conduct sections
Issue #5
