### 2022-07-29 cace15a
Add personal workflow
Issue #5
