### 2022-07-29 6895d52
Add details for Docker deployment
Issue #8
