### 2022-07-29 9dbc08a
Fix search bar; add acknowledgment
Issue #4
