### 2022-07-29 75a01b6
Minor edits
Issue #7
