### 2022-08-03 #4530 @btmonier
> Opened issue [maize-genetics/rTASSEL#3](https://github.com/maize-genetics/rTASSEL/issues/3)

@tomsing1 - I have added new installation instructions found in the following document:

* [README.md](https://github.com/maize-genetics/rTASSEL/blob/master/README.md#installation-and-usage)

I have also created a new installation vignette where I also cover this:
* [rtassel_install.Rmd](https://github.com/maize-genetics/rTASSEL/blob/master/vignettes/rtassel_installation.Rmd)
