### 2022-08-03 #4530 @btmonier
> Opened issue [maize-genetics/rTASSEL#4](https://github.com/maize-genetics/rTASSEL/issues/4)

@tomsing1 - I have updated the `pkgdown` build process in which I have:

* Fixed the article ordering
* Added ORCIDs for each author
* Added links to the GitHub repository
* Added search bar for enhanced navigation (i.e. taking advantage of Bootstrap 5)

Please see the [_pkgdown.yml](https://github.com/maize-genetics/rTASSEL/blob/master/_pkgdown.yml) for further details.
