### 2022-08-03 #4530 @btmonier
> Opened issue [maize-genetics/rTASSEL#5](https://github.com/maize-genetics/rTASSEL/issues/5)

@tomsing1 - I have added the following documentation to help with contribution and also a code of conduct guide as laid out by the Contributor Covenant Code of Conduct.

Please see the following files for further details:

* [CODE_OF_CONDUCT.md](https://github.com/maize-genetics/rTASSEL/blob/master/.github/CODE_OF_CONDUCT.md)
* [CONTRIBUTING.md](https://github.com/maize-genetics/rTASSEL/blob/master/.github/CONTRIBUTING.md)

I have also added links to these files on the right side bar of our [website](https://maize-genetics.github.io/rTASSEL/).
