### 2022-08-03 #4530 @btmonier
> [maize-genetics/rTASSEL#7](https://github.com/maize-genetics/rTASSEL/issues/7)

@tkchafin - I have created a [containerized version of rTASSEL](https://hub.docker.com/repository/docker/maizegenetics/rtassel) based off the [rocker/verse](https://hub.docker.com/r/rocker/verse) Docker image.

Please see the following file for further details:
* [rtassel_installation.Rmd](https://github.com/maize-genetics/rTASSEL/blob/master/vignettes/rtassel_installation.Rmd)
