### 2022-08-03 #4530 @btmonier
> [maize-genetics/rTASSEL#6](https://github.com/maize-genetics/rTASSEL/issues/6)

@tkchafin - We have identified the `log4j` issue and submitted a fix which will be released in the next TASSEL standalone build (2022-08-04). Once this goes into production, we will update the JAR files for the deployment of rTASSEL. Thanks!
