### 2022-08-08 #4530 @fboehm
@btmonier - the next steps are for you to make a new release of rTASSEL and to deposit that new release as an archive, for example, with zenodo.org. Please report here the new release's version number and the doi for the archive. Thank you!
