+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   rtassel_benchmarks: rtassel_benchmarks.html
   rtassel_installation: rtassel_installation.html
   rtassel_walkthrough: rtassel_walkthrough.html
-last_built: 2022-08-03T23:57Z
+last_built: 2022-08-09T17:12Z
 urls:
   reference: https://maize-genetics.github.io/rTASSEL/reference
   article: https://maize-genetics.github.io/rTASSEL/articles
