### 2022-08-10 #4530 @danielskatz
@btmonier - As the AEiC on duty currently, I'll finalize your paper publication. I have suggested some small changes in https://github.com/maize-genetics/rTASSEL/pull/9 and https://github.com/maize-genetics/rTASSEL/pull/10 Please merge these, or let me know what you disagree with, then we can proceed to publication
