+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   genotype_filtration: genotype_filtration.html
   rtassel_benchmarks: rtassel_benchmarks.html
   rtassel_walkthrough: rtassel_walkthrough.html
-last_built: 2022-08-10T01:52Z
+last_built: 2022-08-10T01:57Z
 
